Keep the multi-model AutoTP regression tests inside tests/unit/v1
The full model_parallelism directory stays out of tests/unit/v1 until the
transformers-main 'embedding_rowwise' drift (#8290) is handled, but the two
multi-model regressions this PR adds (#8231 reproduction and the multimodal
text_config path) are safe there and belong to the GPU workflow's scope.
